+++ CalcExpProxy/ICalculator.cs
@@ -4,7 +4,6 @@ namespace CalcExpProxy
 {
     public interface ICalculatorAsync
     {
-        Task<double> CalculateAsync(string expression);
-        Task<double?> GetReqAsync(string expression);
+        Task<double?> CalculateAsync(string expression);
     }
 }
